nvbench_compare.py is now using legacy decision tree
nvbench_compare_robust.py uses new decision tree.
